Extracted 'actionButtons' into separate component
